backend/requirements.txt
groq
python-dotenv
fastapi
uvicorn[standard]
pydantic
